Add ability to lock currently active nessus scans to prevent duplicates
There were instances when a write command from fsnotify was triggered multiple times for the same scan UUID, causing the same scan to be sent multiple times. This adds a concurrently safe list of current nessus scans that keeps track of the currently in process scans to prevent duplicate scans from occurring.

Also broke down some of the functionality into separate functions with extra comments to make the code a little more readable.